+++ tests/frontend/org/voltdb/regressionsuites/failureprocs/CleanupFail.java
@@ -32,7 +32,7 @@ import org.voltdb.*;
 public class CleanupFail extends VoltProcedure {
     public final SQLStmt insert = new SQLStmt("INSERT INTO NEW_ORDER VALUES (?, ?, ?);");
 
-    public long run(long no_o_id, long no_d_id, long no_w_id) {
+    public long run(long no_o_id, long no_d_id, byte no_w_id) {
         voltQueueSQL(insert, no_o_id, no_d_id, no_w_id);
         long count = voltExecuteSQL()[0].asScalarLong();
         voltQueueSQL(insert, no_o_id + 1, no_d_id + 1, no_w_id + 1);
